+++ src/app/Components/donaciones/donaciones/donaciones.component.css
@@ -9,20 +9,11 @@ a:active,
 a:link,
 a:visited {
     text-decoration: none
-}
+} 
 
 body {
-    background-color: #ffffff;
-
-    padding: 0;
-
-    background-size: cover;
-    color: black;
-    font-family: "open sans";
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    font-family: 'Lato', sans-serif;
+    background-color: #fff;
+    margin: 8px 20px
 }
 
 h1 {
@@ -73,11 +64,9 @@ a:link {
     color: #3770B1
 }
 
-/*
 a:visited {
-     color: #57A64C  
+    color: #57A64C
 }
-*/
 
 a:hover {
     color: #4D5FA6
@@ -103,38 +92,6 @@ a:active {
     text-decoration: none;
     text-shadow: 0px -1px 0px #2b665e;
 }
-
-/*header*/
-
-@import url(https://fonts.googleapis.com/css?family=Raleway);
-
-nav {
-    margin-top: 0px;
-    padding: 24px;
-    text-align: center;
-    font-family: Raleway;
-  }
-  
-  #nav-3 {
-    background: var(--blue);
-  }
-  
-  .link-3 {
-    transition: 0.4s;
-    color: #ffffff;
-    font-size: 20px;
-    text-decoration: none;
-    padding: 0 10px;
-    margin: 0 10px;
-  }
-  
-  .link-3:hover {
-    background-color: #ffffff;
-    color: #003cff;
-    padding: 24px 10px;
-    border-radius: 50%;
-  }
-
 
 .indexbutton:hover {
     color: #585858 !important;
@@ -173,7 +130,7 @@ mark.gold {
 mark.bronze {
     font-size: 22px;
     font-weight: bold;
-    color: #ffffff;
+    color: #cd7f32;
     background: none;
 }
 
@@ -184,61 +141,4 @@ mark.silver {
     font-style: italic;
     color: #32DC24;
     background: none;
-}
-
-/*Acordeon desplegable*/
-@import url('https://fonts.googleapis.com/css2?family=Raleway:wght@400;800&display=swap');
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Raleway', sans-serif;
-}
-.contenedor-acordeon{
-    width: 100%;
-    max-width: 800px;
-    margin: auto;
-    margin-top: 80px;
-}
-.contenedor-acordeon h2{
-    text-align: center;
-    font-size: 40px;
-    margin-bottom: 35px;
-}
-.acordeon{
-    background: #fff;
-}
-.acordeon label{
-    display: block;
-    padding: 20px;
-    font-size: 20px;
-    background: #48d6e9;
-    color: #fff;
-    cursor: pointer;
-    margin-bottom: 2px;
-    transition: all 300ms ease;
-}
-.acordeon label:hover{
-    background: #b6f3ff;
-}
-.acordeon .contenido-acordeon{
-    padding: 0px;
-    margin: 0px 20px;
-    max-height: 0px;
-    overflow: hidden;
-    transition: all 300ms ease;
-}
-.btn-acordeon:checked ~ .contenido-acordeon{
-    max-height: 600px;
-    padding: 15px 0px;
-}
-.btn-acordeon{
-    display: none;
-}
-
-
-/*Carruesel*/
-.carousel-item{
-    height: 30rem;
-    background-color: #48d6e9;
 }
