--- src/app/Components/footer/footer.component.css
@@ -1,6 +1,0 @@
-footer {
-    background-color: var(--blue);
-    color: #fff;
-    text-align: center;
-    padding: 30px 0;
-}
