--- src/app/Components/footer/footer.component.html
@@ -1,15 +1,0 @@
-<footer class="footer">
-  
-    <div class="p-3 flex justify-content-between lg:justify-content-center align-items-center flex-wrap">
-      <div class="font-bold mr-8">{{fundacion.nombreFundacion}}</div>
-      <div class="align-items-center hidden lg:flex">
-          <span class="line-height-3">{{fundacion.direccion}}</span>
-      </div>
-      <a class="flex align-items-center ml-2 mr-8">
-          <a class="text-white" href="{{fundacion.paginaWeb}}" ><span class="underline font-bold">IR</span></a>
-      </a>
-      <span pRipple class="flex align-items-center justify-content-center" >
-          <i class="pi pi-clock"> {{fundacion.horarios}}</i>
-      </span>
-      </div>
-    </footer>
