+++ src/app/Components/formAdopcion/for-adopcion/for-adopcion.component.css
@@ -8,7 +8,7 @@
 } */
 
 img {
-  max-width: 100%; /* Hace que la imagen sea responsive */
+
   border-radius: 10px;
 }
 
@@ -120,6 +120,8 @@ label.inputLabel {
   background-color: #5592a7;
 } */
 
+
+
 .layout2 {
   width: 90%;
   height: 100%;
@@ -149,15 +151,10 @@ label.inputLabel {
   width: 100%;
   height: 100%;
 
+  display: grid;
   grid-template-rows: 1fr;
   grid-template-columns: repeat(2, 1fr);
   gap: 8px;
-
-  display: grid;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
 }
 
 
@@ -425,17 +422,4 @@ button:hover .button-text {
 
   flex-direction: column;
   align-items: center;
-}
-
-@media screen and (max-width: 768px){
-  nav {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center; /* Centra verticalmente los enlaces */
-    padding: 10px; /* Ajusta el relleno para dispositivos pequeños */
-    background: var(--blue);
-  }
-  .link-3{
-    margin: 5px;
-  }
 }
