+++ src/app/Components/formAdopcion/for-adopcion/for-adopcion.component.html
@@ -1,24 +1,24 @@
-<app-header></app-header>
 <div class="header" style="width: 100%; height:1000px; background-color: aquamarine;">
 
 
-    
-        <section class="requerimientosAdicionales">
-            <div style="  display: flex;
-      justify-content: center;
-      align-items: center;
-      height: 100%;">
-                <img src="assets/img/logo.png" width="500px" alt="">
+    <section class="requerimientosAdicionales">
+        <div style="  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;">
+            <img src="assets/img/logo.png" width="500px" alt="">
+
+
+        </div>
+        <div>
+            <div class="title">
+                <h1> NO COMPRES, ¡ADOPTA Y SALVA
+                    UNA VIDA!
+                </h1>
             </div>
-            <div>
-                <div class="title">
-                    <h1>NO COMPRES, ¡ADOPTA Y SALVA
-                        UNA VIDA!
-                    </h1>
-                </div>
-            </div>
-        </section>
-    
+        </div>
+    </section>
+</div>
 
 <section class="importantes" style="width: 90%; margin-left: auto; margin-right: auto;">
 
@@ -157,7 +157,6 @@
                                 required />
                             <label for="float-input">Dirección de trabajo</label>
                         </span>
-                        <small class="p-error" *ngIf="submitted && !formulario.direccionTrabajo">{{error}}</small>
                     </div>
 
                     <div class="col-12 md:col-4 mt-2">
@@ -190,7 +189,7 @@
                                 onkeypress='return (event.charCode >= 48 && event.charCode <= 57)' />
                             <label for="float-input">Teléfono Trabajo</label>
                         </span>
-                        <small class="p-error" *ngIf="submitted && !formulario.telefonoTrabajo">{{error}}</small>
+
                     </div>
 
                     <div class="col-12 md:col-4 mt-2">
@@ -378,8 +377,7 @@
                 <!-- 7 -->
                 <div class="p-grid p-fluid mt-2">
                     <div class="p-col-12 p-md-12 mt-2">
-                        <label class="font-bold block mb-2"
-                            [style]="submitted && !formulario.animalPasado7 ? 'color: red':''">
+                        <label class="font-bold block mb-2">
                             7. ¿Qué animal era?
                         </label>
 
@@ -392,9 +390,7 @@
                 <!-- 8 -->
                 <div class="p-grid p-fluid mt-2">
                     <div class="p-col-12 p-md-12 mt-2">
-                        <label class="font-bold block mb-2"
-                            [style]="submitted && !formulario.tiempoVivido8 ? 'color: red':''">
-
+                        <label class="font-bold block mb-2">
                             8. ¿Cuánto tiempo vivió con ustedes?
                         </label>
 
@@ -408,8 +404,7 @@
                 <div class="p-grid p-fluid mt-2">
                     <div class="p-col-12 p-md-12 mt-2">
 
-                        <label class="font-bold block mb-2"
-                            [style]="submitted && !formulario.animalAhora9 ? 'color: red':''">
+                        <label class="font-bold block mb-2">
                             9. ¿En dónde está ahora el animal? ¿Por qué? Falleció, Como?
                         </label>
                         <textarea rows="4" cols="30" pInputTextarea [(ngModel)]="formulario.animalAhora9"
@@ -421,10 +416,10 @@
                 <div class="p-grid p-fluid mt-2">
                     <div class="p-col-12 p-md-12 mt-2">
 
-                        <label class="font-bold block mb-2"
-                            [style]="submitted && !formulario.vacunado10 ? 'color: red':''">
+                        <label class="font-bold block mb-2">
                             10. ¿Estaba vacunado?
                         </label>
+
                         <div class="flex align-items-center">
                             <p-radioButton name="vacunado10" value="SI"
                                 [(ngModel)]="formulario.vacunado10"></p-radioButton>
@@ -445,8 +440,7 @@
 
 
                     <div class="col-12 md:col-6">
-                        <label class="font-bold block mb-2"
-                            [style]="submitted && !formulario.vacunasAntes11 ? 'color: red':''">
+                        <label class="font-bold block mb-2">
                             11. ¿Qué vacunas se le aplicaron? y ¿Cada cuánto tiempo se vacunaba?
                         </label>
                         <textarea rows="4" cols="30" pInputTextarea [(ngModel)]="formulario.vacunasAntes11"
@@ -454,8 +448,7 @@
                     </div>
 
                     <div class="col-12 md:col-6">
-                        <label class="font-bold block mb-2"
-                            [style]="submitted && !formulario.desparasitado12 ? 'color: red':''">
+                        <label class="font-bold block mb-2">
                             12. ¿Estaba desparasitado? ¿Cada cuánto tiempo se desparasitaba?
                         </label>
                         <textarea rows="4" cols="30" pInputTextarea [(ngModel)]="formulario.desparasitado12"
@@ -468,8 +461,7 @@
                 <!-- 13 -->
                 <div class="p-grid p-fluid mt-2">
                     <div class="p-col-12 p-md-12 mt-2">
-                        <label class="font-bold block mb-2"
-                            [style]="submitted && !formulario.horasSolo14 ? 'color: red':''">
+                        <label class="font-bold block mb-2">
                             13. ¿Cuántas horas del día permanecería solo?
                         </label>
                         <p-inputNumber [(ngModel)]="formulario.horasSolo14" name="horasSolo14" mode="decimal"
