--- src/app/Components/header/header.component.css
@@ -1,28 +1,0 @@
-@import url(https://fonts.googleapis.com/css?family=Raleway);
-
-nav {
-  margin-top: 0px;
-  padding: 24px;
-  text-align: center;
-  font-family: Raleway;
-}
-
-#nav-3 {
-  background: var(--blue);
-}
-
-.link-3 {
-  transition: 0.4s;
-  color: #ffffff;
-  font-size: 20px;
-  text-decoration: none;
-  padding: 0 10px;
-  margin: 0 10px;
-}
-
-.link-3:hover {
-  background-color: #ffffff;
-  color: #003cff;
-  padding: 24px 10px;
-  border-radius: 50%;
-}
