--- src/app/Components/header/header.component.html
@@ -1,17 +1,0 @@
-<div class="p-3 flex justify-content-between lg:justify-content-center align-items-center flex-wrap"
-  style="background-color: var(--blue)">
-  <div class="font-bold mr-4">FAAN</div>
-  <div class="align-items-center hidden lg:flex">
-    <span class="line-height-3">🐶 Fundación Familia Amor Animal</span>
-  </div>
-</div>
-<div>
-  <nav id="nav-3">
-    <a class="link-3 " [routerLink]="['/home']" >Home</a>
-    <a class="link-3" [routerLink]="['/informacion']">Información</a>
-    <a class="link-3" [routerLink]="['/formulario-adopcion']">Formulario para adopción</a>
-    <a class="link-3" [routerLink]="['/donaciones']">Donaciones</a>
-    <a class="link-3" [routerLink]="['/contacto']">Contactos</a>
-  </nav>
-</div>
-
