+++ src/app/Components/home/home.component.css
@@ -73,6 +73,14 @@
            padding: 30px 0;
        }
 
+       footer a {
+           text-decoration: none;
+           color: #fff;
+           display: inline-block;
+           margin: 5px;
+           font-size: 26px;
+       }
+
        /* *********************/
        /* seccion RESPONSIVE */
        /* *********************/
@@ -151,40 +159,3 @@
            z-index: 100;
            transition: all 300ms ease;
        }
-
-       /*header*/
-
-@import url(https://fonts.googleapis.com/css?family=Raleway);
-
-.login-link {
-    float: right;
-  }
-
-nav {
-  margin-top: 0px;
-  padding: 24px;
-  text-align: center;
-  font-family: Raleway;
-
-}
-
-#nav-3 {
-  background: var(--blue);
-}
-
-.link-3 {
-  transition: 0.4s;
-  color: #ffffff;
-  font-size: 20px;
-  text-decoration: none;
-  padding: 0 10px;
-  margin: 0 10px;
-}
-
-.link-3:hover {
-  background-color: #ffffff;
-  color: #003cff;
-  padding: 24px 10px;
-  border-radius: 50%;
-}
-
