--- src/app/Components/informacion/informacion/informacion.component.css
@@ -1,204 +1,0 @@
-/*header*/
-
-@import url(https://fonts.googleapis.com/css?family=Raleway);
-
-nav {
-  margin-top: 0px;
-  padding: 24px;
-  text-align: center;
-  font-family: Raleway;
-}
-
-#nav-3 {
-  background: var(--blue);
-}
-
-.link-3 {
-  transition: 0.4s;
-  color: #ffffff;
-  font-size: 20px;
-  text-decoration: none;
-  padding: 0 10px;
-  margin: 0 10px;
-}
-
-.link-3:hover {
-  background-color: #ffffff;
-  color: #003cff;
-  padding: 24px 10px;
-  border-radius: 50%;
-}
-
-.center-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh; /* Asegura que el contenedor ocupe al menos el 100% del alto de la ventana */
-}
-
-.floating-image {
-  max-width: 100%;
-  height: auto;
-  animation: float 2s ease-in-out infinite; /* Aplica la animación */
-}
-
-@keyframes float {
-  0%, 100% {
-    transform: translateY(0); /* Empieza y termina en la posición original */
-  }
-  50% {
-    transform: translateY(-10px); /* Punto más alto del movimiento */
-  }
-}
-
-h1 {
-  font-size: 36px;
-  font-family: Arial, Verdana;
-  color: #70b8d0
-}
-
-body {
-  margin: 0;
-  padding: 0;
-  background-color: #ffffff;
-  background-size: cover;
-  color: black;
-  font-family: "open sans";
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  font-family: 'Lato', sans-serif; /* Cambiar tipo de letra */
-
-}
-
-/* cARRUSEL */
-
-.slider-container {
-  display: flex;
-  width: 100%;
-  height: 100vh;
-  overflow-x: scroll;
-}
-
-.slider-container img {
-  flex: 0 0 100%;
-  width: 100%;
-  object-fit: cover;
-}
-
-/*Informacion*/
-.container-boxes {
-  margin: 0 auto;
-  padding: 0;
-  display: flex;
-  justify-content: space-around;
-  align-items: flex-end;
-  max-width: 90vw;
-}
-
-.box {
-  background: rgba(247, 250, 251);
-  margin: 1rem;
-  padding: .5rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  max-width: 580px;
-  max-height: 580px;
-  min-height: 580px;
-  transition: all 250ms ease-out;
-}
-
-.box:hover {
-  background: rgb(247, 250, 251);
-  transform: translateY(-20%);
-}
-
-.box a {
-  text-decoration: none;
-  color: black;
-}
-
-.icon {
-  font-size: 3rem;
-  padding: 1rem;
-}
-
-.text h3 {
-  text-transform: uppercase;
-  letter-spacing: 4px;
-  margin-bottom: 0;
-}
-
-.text p {
-  margin-top: 2rem;
-  line-height: 1.5rem;
-  text-align: left;
-}
-
-.text {
-  padding: .4rem;
-}
-
-.container-boxes {
-  display: flex;
-  flex-wrap: wrap; /* Permite que los boxes se envuelvan en pantallas más pequeñas */
-  justify-content: space-around;
-  align-items: flex-start; /* Cambiado a "flex-start" para que los boxes se alineen en la parte superior */
-  max-width: 90vw;
-  margin: 20px auto;
-}
-
-.box {
-  background: rgba(247, 250, 251);
-  margin: 1rem;
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  max-width: 380px;
-  transition: transform 0.3s ease;
-  width: calc(33.33% - 2rem); /* Para que haya 3 boxes por fila en pantallas grandes */
-}
-
-.box:hover {
-  transform: translateY(-10px);
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
-}
-
-.text {
-  text-align: center;
-}
-
-.text h3 {
-  text-transform: uppercase;
-  letter-spacing: 2px;
-  margin-bottom: 10px;
-  font-size: 20px;
-}
-
-.text p {
-  margin-top: 0;
-  line-height: 1.5;
-}
-
-.text a {
-  text-decoration: none;
-  color: #0078d4;
-  transition: color 0.3s ease;
-}
-
-.text a:hover {
-  color: #003cff;
-}
-
-/* Media query para pantallas más pequeñas */
-@media (max-width: 768px) {
-  .container-boxes {
-    justify-content: center; /* Centra los boxes en pantallas pequeñas */
-  }
-
-  .box {
-    width: calc(50% - 2rem); /* Para que haya 2 boxes por fila en pantallas medianas */
-  }
-}
