--- src/app/Components/informacion/informacion/informacion.component.html
@@ -1,96 +1,0 @@
-<header>
-  <div class="p-3 flex justify-content-between lg:justify-content-center align-items-center flex-wrap"
-    style="background-color: var(--blue)">
-    <div class="font-bold mr-4">FAAN</div>
-    <div class="align-items-center hidden lg:flex">
-      <span class="line-height-3">🐶 Fundación Familia Amor Animal</span>
-    </div>
-  </div>
-  <nav id="nav-3">
-    <a class="link-3" [routerLink]="['/home']">Home</a>
-    <a class="link-3" [routerLink]="['/informacion']">Informacion</a>
-    <a class="link-3" [routerLink]="['/formulario-adopcion']">Formulario para adopcion</a>
-    <a class="link-3" [routerLink]="['/donaciones']">Donaciones</a>
-    <a class="link-3" [routerLink]="['/contacto']">Contactos</a>
-  </nav>
-</header>
-
-<body style="border-width: 10px; align-items: center;">
-  <h1 style="text-align: center; margin-top: 20px;"> FUNDACION AMOR ANIMAL</h1>
-
-
-  <section>
-    <br>
-    <h3>ENCUENTRANOS EN:</h3>
-
-    <h5>Tarqui (sector estadio de Tarqui), Cuenca, Ecuador</h5>
-    <br>
-    <div class="content">
-      <iframe
-        src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3709.3433028264044!2d-79.03697322536023!3d-3.013359289952116!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x91cd203a6e623497%3A0xf3f7f62bbb1c256b!2sEstadio%20Parroquial%20-Tarqui!5e1!3m2!1ses!2sec!4v1694054160823!5m2!1ses!2sec"
-        width="600" height="450" style="border:0;" allowfullscreen="" loading="lazy"
-        referrerpolicy="no-referrer-when-downgrade"></iframe>
-      <br>
-
-      <br>
-    </div>
-    <div class="content">
-    </div>
-  </section>
-
-  <br>
-  <br>
-  <h3>ACERCA DE NOSOTROS:</h3>
-  <br>
-  <br>
-
-  <section class="container-boxes">
-    <div class="box">
-      <div class="text">
-        <a >
-          <h3>¿QUIENES SOMOS?</h3>
-          <p>FAAN es un grupo de defensa animal, sin ánimo de lucro, que tiene como finalidad ayudar y acoger a
-            cuantos animales abandonados en situaciones de riesgo se pueda, a su vez busca crear conciencia en cuanto
-            al respeto animal.</p>
-        </a>
-      </div>
-
-    </div>
-
-    <div class="box">
-
-
-
-      <div class="text">
-        <a >
-          <h3>OBJETIVOS</h3>
-          <p>
-            • Rescatar animales abandonados en situación de riesgo y buscarles un hogar responsable luego de haber
-            tenido atención médica oportuna y rehabilitación social en caso de requerirse.
-            <br>• Disminuir el maltrato animal creando conciencia acerca del tema.
-            <br>• Fomentar la importancia de la esterilización de animales de compañía para evitar la sobrepoblación.
-            <br>• Reducir el número de animales abandonados
-          </p>
-        </a>
-      </div>
-
-    </div>
-
-    <div class="box">
-
-
-
-      <div class="text">
-        <a >
-          <h3>FINANCIAMIENTO
-          </h3>
-          <p>Nos financiamos de las donaciones realizadas por personas particulares o empresas que desean aportar con
-            nuestra causa , además de recaudar fondos mediante diversas actividades realizadas por el grupo , tales
-            como ventas de artículos donados, realización de rifas, eventos y venta de productos varios.</p>
-        </a>
-      </div>
-
-    </div>
-  </section>
-
-</body>
