--- src/app/Components/informacion/informacion/informacion.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { InformacionComponent } from './informacion.component';
-
-describe('InformacionComponent', () => {
-  let component: InformacionComponent;
-  let fixture: ComponentFixture<InformacionComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [InformacionComponent]
-    });
-    fixture = TestBed.createComponent(InformacionComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
