--- src/app/Components/informacion/informacion/informacion.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-informacion',
-  templateUrl: './informacion.component.html',
-  styleUrls: ['./informacion.component.css']
-})
-export class InformacionComponent {
-
-}
