+++ src/app/Modules/Administrador/animal/register-tipo-animal/register-tipo-animal.component.html
@@ -24,7 +24,7 @@
             <div class="grid p-fluid align-items-center">
                 <div class="col-12 md:col-8">
                     <div class="p-inputgroup">
-                        <h5 class="m-0"><strong>Tipos de animales</strong></h5><br>
+                        <h5 class="m-0"><strong>Tipos animales</strong></h5><br>
                     </div>
                 </div>
