+++ src/app/Modules/Administrador/control-animal/control-animal.component.ts
@@ -70,7 +70,7 @@ export class ControlAnimalComponent implements OnInit {
 
 
 
-  isTextDigit: string = '';
+  isTextDigit!: string;
   public getAllMascotas(): void {
     try {
       this.animalesService.getAllAnimalesPagesOrPlacaOrName(this.isTextDigit, this.isPage, this.isSize, this.isSosrt).subscribe((data: any) => {
@@ -210,10 +210,9 @@ export class ControlAnimalComponent implements OnInit {
   saveTipoVacuna() {
     this.tipoVacuna.estado = true;
     this.tipoVacunaService.saveTipoVacuna(this.tipoVacuna).subscribe((data) => {
-     
+      alert('SUCESSFULL');
       this.tipoVacuna = {} as TipoVacuna;
       this.visibleTipoVacuna = false;
-      this.getAllTiposVacunas();
     })
   }
 
@@ -242,6 +241,7 @@ export class ControlAnimalComponent implements OnInit {
   saveEstadoAnimal() {
     this.estadoAnimal.estado = 'A';
     this.estadoAnimalService.saveEstadoAnimal(this.estadoAnimal).subscribe(data => {
+      alert('SUCESSFULL');
       this.getListEstadoAnimal();
     })
   }
@@ -338,6 +338,7 @@ export class ControlAnimalComponent implements OnInit {
       this.vacuna.estadoVacuna = true;
     this.vacunaService.saveVacuna(this.vacuna).subscribe((data) => {
       console.log(data);
+      alert('SUCESSFULL');
       this.getListaVacunasByIdControlAnimal(this.isControlAnimal.idControlAnimal!)
       this.vacuna = {} as Vacuna;
       this.isControlAnimal = {} as ControlAnimal;
