+++ src/app/Modules/Administrador/control-person/control-person.component.html
@@ -21,7 +21,7 @@
             <div class="grid p-fluid align-items-center">
                 <div class="col-12 md:col-8">
                     <div class="p-inputgroup">
-                        <h5 class="m-0"><strong>Gestión de las Personas</strong></h5><br>
+                        <h5 class="m-0"><strong>Gestión Personas</strong></h5><br>
                     </div>
                 </div>
 
@@ -100,10 +100,10 @@
                 formcontrolname="idPersona" />
 
             <div class="field">
-                <label for="identificacion"><strong>Ingrese su identificacion</strong></label>
+                <label for="identificacion"><strong>Ingrese su identificacion electrónico.</strong></label>
                 <input type="identificacion" pInputText id="identificacion" name="identificacion"
-                    placeholder="Cédula..." formControlName="identificacion"
-                    [class.is-invalid]="(respose && respose.ci) || formPerson.get('identificacion')!.invalid && formPerson.get('identificacion')!.touched"
+                    placeholder="Dirección de identificacion electrónico" formControlName="identificacion"
+                    [class.is-invalid]="formPerson.get('identificacion')!.invalid && formPerson.get('identificacion')!.touched"
                     class="form-control" formcontrolname="identificacion" [maxlength]="10"
                     onkeypress='return event.charCode >= 48 && event.charCode <= 57' required autofocus />
 
@@ -115,10 +115,6 @@
                     <small class="text-danger"
                         *ngIf="formPerson.get('identificacion')?.errors?.['minlength']"><strong>Minimo
                             requerido es 10.</strong></small>
-                    <small class="text-danger" *ngIf="respose && respose.ci">
-                        <strong>{{respose.ci}}</strong></small>
-
-
                 </div>
 
 
@@ -129,7 +125,7 @@
                 <label for="correo"><strong>Ingrese su correo electrónico.</strong></label>
                 <input type="correo" pInputText id="correo" name="correo" placeholder="Ej: example@email.com"
                     formControlName="correo"
-                    [class.is-invalid]="(respose && respose.emailValidate) || formPerson.get('correo')!.invalid && formPerson.get('correo')!.touched"
+                    [class.is-invalid]="formPerson.get('correo')!.invalid && formPerson.get('correo')!.touched"
                     class="form-control" formcontrolname="correo" required />
 
                 <div
@@ -139,8 +135,6 @@
                     <small class="text-danger" *ngIf="formPerson.get('correo')?.errors?.['pattern']"><strong>Dirección
                             de correo
                             invalido.</strong> </small>
-                    <small class="text-danger" *ngIf="respose && respose.emailValidate">
-                        <strong>{{respose.emailValidate}}</strong></small>
                 </div>
             </div>
