+++ src/app/Modules/Administrador/control-person/control-person.component.ts
@@ -5,11 +5,9 @@ import {
   FormGroup,
   Validators,
 } from '@angular/forms';
-import { forkJoin } from 'rxjs';
 import { Persona } from 'src/app/Models/persona';
 import { PersonaService } from 'src/app/Service/persona.service';
 import { ScreenSizeService } from 'src/app/Service/screen-size-service.service';
-import { Response, ValidateEquals } from 'src/app/util/model/response-validate';
 
 @Component({
   selector: 'app-control-person',
@@ -19,15 +17,14 @@ import { Response, ValidateEquals } from 'src/app/util/model/response-validate';
 export class ControlPersonComponent implements OnInit {
   public personDialog: boolean = false;
 
-  public respose = new Response();
-
   public person = new Persona();
 
   public submitted: boolean = false;
 
   public listPerson: Persona[] = [];
 
-  public validateEquals = new ValidateEquals();
+  public errorUnique: string = '';
+
   //Size of window..
   public screenWidth: number = 0;
   public screenHeight: number = 0;
@@ -103,9 +100,9 @@ export class ControlPersonComponent implements OnInit {
 
     return new Date().getFullYear() - fechaNacimiento.getFullYear() < 18
       ? {
-        fechaNacimientoInvalida: true,
-        mensaje: 'Debe ser mayor de edad.',
-      }
+          fechaNacimientoInvalida: true,
+          mensaje: 'Debe ser mayor de edad.',
+        }
       : null;
   }
 
@@ -175,36 +172,14 @@ export class ControlPersonComponent implements OnInit {
       this.formPerson.markAllAsTouched();
       return;
     }
-
     this.submitted = true;
     this.person = this.formPerson.value;
 
-    this.validateEmailAndCi(this.person.identificacion!, this.person.correo!);
-  }
-
-  public validateEmailAndCi(identificacion: string, email: string) {
-
-    const identificacionEqualsValidate = this.validateEquals.identificacion === this.person.identificacion;
-    const emailEqualsValidate = this.validateEquals.emailValidate === this.person.correo;
-
-    const existIdentificacion = this.personService.existByIdentificacion(identificacion, identificacionEqualsValidate);
-    const existEmail = this.personService.existsByEmail(email, emailEqualsValidate);
-
-    forkJoin([existIdentificacion, existEmail]).subscribe(
-      ([cedulaRepetidaResp, emailRepetidoResp]) => {
-
-        this.respose.ci = cedulaRepetidaResp ? 'Identificación existente' : ''
-        this.respose.emailValidate = emailRepetidoResp ? 'Direccón de correo existente' : '';
-
-        if (!cedulaRepetidaResp && !emailRepetidoResp) {
-          if (this.person.idPersona) {
-            this.updatePerson();
-          } else {
-            this.savePerson();
-          }
-        }
-      }
-    );
+    if (this.person.idPersona) {
+      this.updatePerson();
+    } else {
+      this.savePerson();
+    }
   }
 
   public isEmpty(obj: any) {
@@ -219,7 +194,9 @@ export class ControlPersonComponent implements OnInit {
         this.closeDialog();
       },
       error: (err) => {
-        alert('err');
+        if (err.status === 400) {
+          this.errorUnique = 'Nombre existente.';
+        }
       },
     });
   }
@@ -229,17 +206,24 @@ export class ControlPersonComponent implements OnInit {
       .updatePersona(this.person.idPersona!, this.person)
       .subscribe({
         next: (resp) => {
-          alert('succesfull updated..');
-          const indexfind = this.listPerson.findIndex(
-            (person) => person.idPersona === resp.idPersona
-          );
-          this.listPerson[indexfind] = resp;
-          this.closeDialog();
-
+          if (resp != null) {
+            try {
+              const indexfind = this.listPerson.findIndex(
+                (person) => person.idPersona === resp.idPersona
+              );
+              this.listPerson[indexfind] = resp;
+            } catch (error) {
+              throw new Error();
+            }
+            this.closeDialog();
+            alert('succesfull updated..');
+          }
         },
         error: (err) => {
           console.log(err);
-          alert('err');
+          if (err.status === 400) {
+            this.errorUnique = 'Nombre existente.';
+          }
         },
       });
   }
@@ -253,22 +237,21 @@ export class ControlPersonComponent implements OnInit {
   public closeDialog(): void {
     this.personDialog = false;
     this.person = {} as Persona;
+    this.errorUnique = '';
   }
 
   public openNewPerson() {
-    this.respose = {} as Response;
-    this.validateEquals = {} as ValidateEquals;
+    this.errorUnique = '';
     this.formPerson.reset();
     this.person = {} as Persona;
     this.submitted = false;
     this.personDialog = true;
   }
 
-
   public editPerson(person: Persona) {
+    this.errorUnique = '';
     this.formPerson.reset();
     this.person = { ...person };
-    this.validateEquals = { identificacion: this.person.identificacion, emailValidate: this.person.correo }
     this.person.fechaNacimiento = new Date(this.person.fechaNacimiento!);
     this.formPerson.patchValue(this.person);
     this.personDialog = true;
