+++ src/app/Modules/Administrador/registro-mascotas/registro-mascotas.component.html
@@ -69,8 +69,7 @@
 			<tr>
 
 				<td>{{ animal.idAnimal }}</td>
-				<td><img [src]="'https://bucketgtmt.s3.amazonaws.com/'+ animal.fotoAnimal" width="100" height="60"
-						class="shadow-6" /></td>
+				<td><img [src]="getUriFile(animal.fotoAnimal)" width="100" height="60" class="shadow-6" /></td>
 
 				<td>{{ animal.nombreAnimal }}</td>
 				<td>{{ animal.placaAnimal }}</td>
@@ -351,7 +350,7 @@
 							<div class="perfil-usuario-header">
 								<div class="perfil-usuario-portada">
 									<div class="perfil-usuario-avatar">
-										<img [src]="avatarURL ? avatarURL: animal.fotoAnimal ? 'https://bucketgtmt.s3.amazonaws.com/' +animal.fotoAnimal: 'https://us.123rf.com/450wm/dervish37/dervish371805/dervish37180500017/101552833-icono-de-vector-gato-y-perro.jpg?ver=6'"
+										<img [src]="avatarURL ? avatarURL: animal.fotoAnimal ? getUriFile(animal.fotoAnimal): 'https://us.123rf.com/450wm/dervish37/dervish371805/dervish37180500017/101552833-icono-de-vector-gato-y-perro.jpg?ver=6'"
 											alt="img-avatar" height="165px">
 										<input type="file" #fileInput style="display: none;"
 											accept="image/jpeg, image/jpg, image/png" (change)="onBasicUpload($event)">
