+++ src/app/Modules/SuperAdministrador/control-usuarios/control-usuarios.component.html
@@ -98,10 +98,7 @@
                 <div class="col-12 md:col-8">
                     <div class="grid p-fluid align-items-center">
                         <div class="col-12 md:col-4">
-                            <label for="nombreRaza"><strong>Identificación</strong></label>
-                            <!-- <div class="card flex justify-content-center">
-                                <p-autoComplete [(ngModel)]="selectedPeople" [suggestions]="filteredPeople" (completeMethod)="filterPeople($event)" field="name"></p-autoComplete>
-                            </div> -->
+                            <label for="nombreRaza"><strong>Identificacion</strong></label>
                             <span class="p-input-icon-right ">
                                 <i class="pi pi-info-circle"
                                     [style.color]="submitted && !persona.identificacion ? 'red':''"></i>
@@ -109,7 +106,7 @@
                                     <input type="text" pTooltip="Identificación" [(ngModel)]="persona.identificacion"
                                         [class]="submitted && (!persona.identificacion) ? 'ng-invalid ng-dirty':''"
                                         id="identificacion" name="identificacion" pInputText
-                                        placeholder="identificacion.." (input)="obtenerNumeroDeLetras()" required />
+                                        placeholder="identificacion.." required />
                                 </div>
                             </span>
                             <small class="p-error" *ngIf="submitted && !persona.identificacion">Identificacion es
@@ -199,42 +196,13 @@
                         </div>
                     </div>
 
-                    <div class="card col-12 md:col-8">
-                        <p class="m-1"><strong>Seleccione los roles:</strong></p>
-
-
-                        <div *ngFor="let role of listRoles" class="field-checkbox mt-2">
-
-
-                            <button *ngIf="isRoleAssigned(role)" type="button" pButton pRipple [label]="role.nombreRol!"
-                                class="w-full custom-button" (click)="addRolsUser(role)"
-                                [ngStyle]="isRoleAssigned(role) ? {color: 'black'}:{color: 'white'}">
-                                <p-tag [icon]="'pi pi-'+(isRoleAssigned(role)? 'times':'check')"
-                                    [severity]="isRoleAssigned(role)? 'success':'danger'"
-                                    [value]="isRoleAssigned(role)?'Rol Asignado':'Asignar Rol'"
-                                    [pTooltip]="'Aplaste el boton para '+(!isRoleAssigned(role)?'Añadir':'Quitar')+' el Rol.'"
-                                    tooltipPosition="bottom"></p-tag>
-                            </button>
-
-
-                            <button *ngIf="!isRoleAssigned(role)" type="button" pButton pRipple
-                                [label]="role.nombreRol!" class="w-full custom-button" (click)="addRolsUser(role)"
-                                [ngStyle]="isRoleAssigned(role) ? {color: 'black'}:{color: 'white'}">
-                                <p-tag [icon]="'pi pi-'+(isRoleAssigned(role)? 'times':'check')"
-                                    [severity]="isRoleAssigned(role)? 'success':'danger'"
-                                    [value]="isRoleAssigned(role)?'Rol Asignado':'Asignar Rol'"
-                                    [pTooltip]="'Aplaste el botón para '+(!isRoleAssigned(role)?'Añadir':'Quitar')+' el Rol.'"
-                                    tooltipPosition="bottom"></p-tag>
-                            </button>
-
+                    <div class="card col-12 md:col-12">
+                        <p class="m-1"><strong>Selecione los roles:</strong></p>
+                        <div *ngFor="let roles of listRoles" class="field-checkbox m-1">
+                            <p-checkbox name="roles" [value]="roles" [(ngModel)]="selectedRoles"
+                                [inputId]="roles.nombreRol"></p-checkbox>
+                            <label [for]="roles.idRol">{{ roles.nombreRol }}</label>
                         </div>
-
-                        <!-- <div *ngFor="let role of listRoles" class="field-checkbox">
-                            <input type="checkbox" [checked]="isRoleAssigned(role)"
-                                (change)="asignarRolesUsuario(role)">
-                            <label>{{ isRoleAssigned(role) ? 'Asig' : 'No asignado' }}</label>
-                        </div> -->
-
                     </div>
 
                 </div>
