+++ src/app/Modules/SuperAdministrador/control-usuarios/control-usuarios.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { AutoCompleteCompleteEvent } from 'primeng/autocomplete';
 import { Persona } from 'src/app/Models/persona';
 import { Rol } from 'src/app/Models/rol';
 import { Usuario } from 'src/app/Models/usuario';
@@ -8,12 +7,16 @@ import { PersonaService } from 'src/app/Service/persona.service';
 import { RolService } from 'src/app/Service/rol.service';
 import { ScreenSizeService } from 'src/app/Service/screen-size-service.service';
 import { UsuarioService } from 'src/app/Service/usuario.service';
+import { FOLDER_IMAGES, getFile } from 'src/app/util/const-data';
 
 @Component({
   selector: 'app-control-usuarios',
   templateUrl: './control-usuarios.component.html',
   styleUrls: ['./control-usuarios.component.css']
 })
+
+
+
 
 export class ControlUsuariosComponent implements OnInit {
 
@@ -26,21 +29,10 @@ export class ControlUsuariosComponent implements OnInit {
 
   ) { }
 
-  numerosIdentificacio?: number;
-
-  obtenerNumeroDeLetras() {
-    this.numerosIdentificacio = this.persona.identificacion!.length;
-    console.log("numero ---> " + this.numerosIdentificacio);
-    if (this.numerosIdentificacio >= 10) {
-      this.pageablePersonaBusqueda();
-    }
-  }
-
   ngOnInit(): void {
     this.getAllUsuario(0, this.size, ['idUsuario', 'asc']);
     this.getSizeWindowResize();
     this.loading = true;
-    this.getAllRolesFull();
   }
 
   //VARIABLE FOR SEARCH BY ATRIBUTE NAME
@@ -65,12 +57,15 @@ export class ControlUsuariosComponent implements OnInit {
 
   public async uploadImage() {
     try {
-      const result = await this.imagenService.savePictureInBuket(this.selectedFile).toPromise();
+      const result = await this.imagenService.savePictureInBuket(this.selectedFile, FOLDER_IMAGES).toPromise();
       return result.key;
     } catch (error) {
       throw new Error()
     }
   }
+
+
+
 
   checked: boolean = true;
 
@@ -92,9 +87,6 @@ export class ControlUsuariosComponent implements OnInit {
     this.usuario = { ...usuario };
     this.persona = this.usuario.persona;
     this.userDialog = true;
-
-    //ADD FOR ME------------------------------------------------------------------------------
-    this.rolsAddUser = [...this.usuario.roles!];
   }
 
   //CHANGE STATE
@@ -126,17 +118,17 @@ export class ControlUsuariosComponent implements OnInit {
   usuario = new Usuario();
 
   public async saveNewUsuario() {
-    ///const key = await this.uploadImage();
+    const key = await this.uploadImage();
     this.personaService.savePersona(this.persona).subscribe((data) => {
       this.persona = data
+      console.log(this.persona)
       this.usuario.idUsuario = 0;
       this.usuario.persona = this.persona;
       this.usuario.roles = this.selectedRoles;
       this.usuario.estadoUsuario = true;
-      this.usuario.fotoPerfil = "key";
+      this.usuario.fotoPerfil = key;
       this.usuarioService.saveUsuario(this.usuario).subscribe((data) => {
         this.usuario = data;
-        this.usuario.roles = this.rolsAddUser; //ADD FOR ME
         alert('SUCESSFULL')
       }, (error) => {
         console.log('2', error)
@@ -149,18 +141,15 @@ export class ControlUsuariosComponent implements OnInit {
 
   // UPDATE USUARIO
   public updateUsuario() {
-    this.usuario.roles = this.rolsAddUser //ADD FOR ME
+    this.usuario.roles = this.selectedRoles
     this.usuarioService.saveUsuario(this.usuario).subscribe((data) => {
       if (data != null) {
         this.usuario = { ...this.usuario }
-        const index = this.listUsuarios.findIndex(i => i.idUsuario === data.idUsuario)
-        this.listUsuarios[index] = data;
 
         this.personaService.updatePersona(this.persona.idPersona!, this.persona)
           .subscribe((data1) => {
             if (data1 != null) {
               alert('datos actualizados')
-              this.userDialog = false;
             }
           })
       }
@@ -187,7 +176,7 @@ export class ControlUsuariosComponent implements OnInit {
   desactivarUser = false;
 
   public openNewUsuario() {
-    this.rolsAddUser = [];
+    this.getAllRolesFull();
     this.persona = {} as Persona;
     this.usuario = {} as Usuario;
     this.userDialog = true;
@@ -238,7 +227,6 @@ export class ControlUsuariosComponent implements OnInit {
     }
   }
 
-
   public hideDialog() {
     this.desactivarUser = false;
     this.userDialog = false;
@@ -249,54 +237,9 @@ export class ControlUsuariosComponent implements OnInit {
     this.fullname = '';
   }
 
-  // AUTOCOMPLETE
-  listPeople: Persona[] = [];
-  selectedPeople = new Persona();
-  filteredPeople: any;
-  loadingPerson!: boolean;
-  sort: string[] = ["identificacion"];
-
-  public pageablePersonaBusqueda() {
-    this.personaService.getAllPersonasPagesOrCedulaOrApellido(this.persona.identificacion!, 0, 50, this.sort)
-      .subscribe((data: any) => {
-        this.listPeople = data.content;
-        this.loadingPerson = false;
-        this.listPeople.forEach(element => {
-          this.persona = element
-        });
-      });
-  }
-
-  public buscar(identificacion: string) {
-    if (identificacion.length >= 10) {
-      this.pageablePersonaBusqueda();
-    }
-  }
-
-
-  //ADD NEW METHODS -- ROLS ADD--------------------------------------
-  public rolsAddUser: Rol[] = [];
-  //ASIGNAR ROLES A USUARIO
-  public addRolsUser(rol: Rol) {
-
-    const index = this.rolsAddUser.findIndex(
-      (item) => item.idRol === rol.idRol
-    );
-
-    if (index !== -1) {
-      this.rolsAddUser.splice(index, 1);
-    } else {
-      this.rolsAddUser.push(rol);
-    }
-
-    console.log(this.rolsAddUser);
-  }
-
-  public isRoleAssigned(role: Rol): boolean {
-
-    return this.rolsAddUser.some(
-      (assignedRole) => assignedRole.idRol === role.idRol
-    );
+  //OBTENER LA IMAGEN NEW MOTHOD------------------------------
+  public getUriFile(fileName: string): string {
+    return getFile(fileName, FOLDER_IMAGES);
   }
 
 }
