+++ src/app/environment/enviroment.ts
@@ -1,0 +1,6 @@
+export const environment = {
+
+  apiuri: 'http://localhost:8080/api',
+  apiuriPublic: 'http://localhost:8080',
+
+};
