--- src/app/shared/breadcrumbs/breadcrumbs.component.ts
@@ -1,36 +1,0 @@
-import { Component, OnDestroy } from '@angular/core';
-import { ActivationEnd, Router } from '@angular/router';
-import { Subscription } from 'rxjs';
-import { map, filter } from 'rxjs/operators';
-
-
-@Component({
-  selector: 'app-breadcrumbs',
-  templateUrl: './breadcrumbs.component.html',
-  styleUrls: ['./breadcrumbs.component.css']
-})
-export class BreadcrumbsComponent implements OnDestroy {
-
-  public titulo?: string;
-  public tituloSubs$: Subscription;
-
-  constructor(private router: Router) {
-    this.tituloSubs$ = this.getArgumentos().subscribe(({ titulo }) => {
-      this.titulo = titulo;
-      document.title = `AdminLte - ${titulo}`;
-    })
-  }
-
-  ngOnDestroy(): void {
-    throw new Error('Method not implemented.');
-  }
-
-  public getArgumentos() {
-    return this.router.events.pipe(
-      filter((event: any) => event instanceof ActivationEnd),
-      filter((event: ActivationEnd) => event.snapshot.firstChild === null),
-      map((event: ActivationEnd) => event.snapshot.data)
-    );
-
-  }
-}
