--- src/app/shared/footerd/footerd.component.html
@@ -1,8 +1,0 @@
-<footer id="footer" class="footer">
-    <div class="copyright">
-        &copy; Copyright <strong><span>FAAN</span></strong>. Todos los derechos reservados
-    </div>
-    <div class="credits">
-        Designed by me!!
-    </div>
-</footer><!-- End Footer -->
