--- src/app/shared/footerd/footerd.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-footerSB',
-  templateUrl: './footerd.component.html',
-  styleUrls: ['./footerd.component.css']
-})
-export class FooterdComponent {
-
-}
