--- src/app/shared/sidebar/sidebar.component.html
@@ -1,192 +1,0 @@
-<div>
-    <!-- ======= Header ======= -->
-    <header id="header" class="header fixed-top d-flex align-items-center">
-
-        <div class="d-flex align-items-center justify-content-between">
-            <a href="index.html" class="logo d-flex align-items-center">
-                <img src="assets/img/logo.png" alt="">
-                <span class="d-none d-lg-block">FundaciónFAAN</span>
-            </a>
-            <i class="bi bi-list toggle-sidebar-btn"></i>
-        </div><!-- End Logo -->
-
-        <nav class="header-nav ms-auto">
-            <ul class="d-flex align-items-center">
-
-                <li class="nav-item dropdown">
-
-                    <a class="nav-link nav-icon" href="#" data-bs-toggle="dropdown">
-                        <i class="bi bi-bell"></i>
-                        <span class="badge bg-primary badge-number">{{countNotificaciones}}</span>
-                    </a><!-- End Notification Icon -->
-
-                    <ul class="dropdown-menu dropdown-menu-end dropdown-menu-arrow notifications">
-                        <li class="dropdown-header">
-                            En este momento usted tiene {{countNotificaciones}} notifications
-                        </li>
-                        <li>
-                            <hr class="dropdown-divider">
-                        </li>
-                        <!--  -->
-                        <p-scrollPanel [style]="{ width: '100%', maxHeight: '400px' }">
-
-                            <!--  -->
-                            <div class="notification" *ngFor="let data of listNotificaciones">
-                                <li class="notification-item">
-                                    <!-- <i class="bi bi-exclamation-circle text-warning"></i> -->
-                                    <i [class]="'bi bi-exclamation-circle text-'+ dayExact(data.diasFaltantes)"></i>
-                                    <div>
-                                        <h4>{{ data.cuerpoMensaje }}</h4>
-                                        <p>{{ data.fullNameMascota }}</p>
-                                        <p>{{formatDate(data.proximaFechaFacunacion) }}</p>
-
-                                        <div class="flex justify-content mt-3">
-                                            <p-button label="Marcar como visto" icon="pi pi-eye"
-                                                styleClass="p-button-sm"
-                                                (onClick)="removerNotificacion(data)"></p-button>
-
-                                        </div>
-
-                                    </div>
-                                </li>
-
-                                <li>
-                                    <hr class="dropdown-divider">
-                                </li>
-                            </div>
-
-                            <!--  -->
-
-                        </p-scrollPanel>
-                        <li>
-                            <hr class="dropdown-divider">
-                        </li>
-
-
-                    </ul><!-- End Notification Dropdown Items -->
-
-                </li><!-- End Notification Nav -->
-
-                <li class="nav-item dropdown pe-3">
-
-                    <a class="nav-link nav-profile d-flex align-items-center pe-0" href="#" data-bs-toggle="dropdown">
-                        <img src="assets/img/dog_eyes.jpg" alt="Profile" class="rounded-circle" width="30px"
-                            height="30px">
-                        <span class="d-none d-md-block dropdown-toggle ps-2">{{userLoggin.toUpperCase()}}</span>
-                    </a><!-- End Profile Iamge Icon -->
-
-                    <ul class="dropdown-menu dropdown-menu-end dropdown-menu-arrow profile">
-                        <li class="dropdown-header">
-                            <h6>{{userLoggin.toUpperCase()}}</h6>
-                            <span>{{rolLoggin}}</span>
-                        </li>
-                        <li>
-                            <hr class="dropdown-divider">
-                        </li>
-
-                        <li>
-                            <a class="dropdown-item d-flex align-items-center" routerLink="perfil-usuario">
-                                <i class="bi bi-person"></i>
-                                <span>Mi Perfil</span>
-                            </a>
-                        </li>
-                        <li>
-                            <hr class="dropdown-divider">
-                        </li>
-
-                        <li>
-                            <a class="dropdown-item d-flex align-items-center" href="users-profile.html">
-                                <i class="bi bi-gear"></i>
-                                <span>Account Settings</span>
-                            </a>
-                        </li>
-
-                        <li>
-                            <hr class="dropdown-divider">
-                        </li>
-
-                        <li (click)="logOut()">
-                            <a class="dropdown-item d-flex align-items-center">
-                                <i class="bi bi-box-arrow-right"></i>
-                                <span>Sign Out</span>
-                            </a>
-                        </li>
-
-                    </ul><!-- End Profile Dropdown Items -->
-                </li><!-- End Profile Nav -->
-
-            </ul>
-        </nav><!-- End Icons Navigation -->
-
-    </header><!-- End Header -->
-
-    <!-- ======= Sidebar ======= -->
-    <aside id="sidebar" class="sidebar">
-
-        <ul class="sidebar-nav" id="sidebar-nav">
-
-            <li class="nav-item">
-                <a class="nav-link " routerLink="/dashboard">
-                    <i class="bi bi-grid"></i>
-                    <span>Dashboard</span>
-
-                </a>
-            </li>
-            <!-- End Dashboard Nav -->
-
-
-
-            <li class="nav-item" *ngFor="let menuP of listSidebar">
-                <a class="nav-link collapsed" [attr.data-bs-target]="'#' + menuP.idNavigation" data-bs-toggle="collapse"
-                    href="#">
-                    <i [class]="menuP.icon"></i><span>{{menuP.title}}</span><i class="bi bi-chevron-down ms-auto"></i>
-                </a>
-                <ul [id]="menuP.idNavigation" class="nav-content collapse " data-bs-parent="#sidebar-nav">
-                    <li *ngFor="let subMenu of menuP.subMenu">
-                        <a [routerLink]="subMenu.path">
-                            <i [class]="'bi bi-circle'+(isCurrentRoute(subMenu.path) ? '-fill':'')"></i><span
-                                [ngStyle]="isCurrentRoute(subMenu.path) ? {'color': '#4154f1'}:{}">{{subMenu.subTitle}}</span>
-
-                        </a>
-                    </li>
-                </ul>
-            </li>
-
-            <!-- End  Nav NG-->
-
-            <li class="nav-heading">Pages</li>
-
-            <li class="nav-item">
-                <a class="nav-link collapsed" routerLink="fundacion">
-                    <i class="bi bi-building-fill-gear"></i>
-                    <span>Fundacion</span>
-                </a>
-            </li>
-            <!-- End  Page Nav -->
-
-        </ul>
-
-    </aside><!-- End Sidebar-->
-
-    <main id="main" class="main">
-
-        <div class="pagetitle">
-
-            <app-breadcrumbs></app-breadcrumbs>
-
-        </div><!-- End Page Title -->
-
-        <section class="section dashboard">
-            <div class="row">
-
-                <router-outlet></router-outlet>
-
-            </div>
-        </section>
-
-    </main><!-- End #main -->
-
-    <!-- ======= Footer ======= -->
-    <app-footerSB></app-footerSB>
-
-</div>
