--- src/app/util/model/response-validate.ts
@@ -1,9 +1,0 @@
-export class Response {
-    ci?: string;
-    emailValidate?: string;
-}
-
-export class ValidateEquals {
-    identificacion?: string;
-    emailValidate?: string;
-}
